Vagrant Setup Done

+++ server/setup/install.sh
@@ -12,34 +12,48 @@ sudo apt-get update
 sudo apt-get install -y nodejs
 
 ### Install Orace Java 7
+echo debconf shared/accepted-oracle-license-v1-1 select true | sudo debconf-set-selections
+echo debconf shared/accepted-oracle-license-v1-1 seen true | sudo debconf-set-selections
 sudo apt-get install -y oracle-java7-installer
 
 ### Install MySQL
+if [[ $1 == "dev" || $1 == "development" ]]; then
+	sudo debconf-set-selections <<< 'mysql-server mysql-server/root_password password root'
+	sudo debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password root'
+fi
 sudo apt-get install -y mysql-server
-
-### Install Reverse Proxy
-sudo apt-get install -y nginx
 
 ### Install Node Libraris
 sudo npm -g install forever
 sudo npm -g install http-proxy
 
-### Stop Apache
-sudo /etc/init.d/apache2 stop
 
-### Load nginx conf
+# Environment Setup
+if [[ $1 == "dev" || $1 == "development" ]]; then
+	#------ DEVELOPMENT ------#
+	echo "Running DEVELOPMENT"
+else
+	#------ PRODUCTION ------#
+	echo "Running PRODUCTION"
+	
+	### Stop Apache
+	sudo /etc/init.d/apache2 stop
+	
+	### Load nginx 
+	sudo apt-get install -y nginx
 
-# Main nginx conf
-cd /etc/nginx
-sudo rm nginx.conf
-sudo wget https://raw.github.com/kubasub/chinese-checkers/feature/server_setup/server/setup/nginx/codecanister.com.conf
+	# Main nginx conf
+	cd /etc/nginx
+	sudo rm nginx.conf
+	sudo wget https://raw.github.com/kubasub/chinese-checkers/develop/server/setup/nginx/codecanister.com.conf
 
-# Codecanister conf
-cd /etc/nginx/conf.d
-sudo rm codecanister.com.conf
-sudo wget https://raw.github.com/kubasub/chinese-checkers/develop/server/setup/nginx/nginx.conf
+	# Codecanister conf
+	cd /etc/nginx/conf.d
+	sudo rm codecanister.com.conf
+	sudo wget https://raw.github.com/kubasub/chinese-checkers/develop/server/setup/nginx/nginx.conf
 
-# Restart
-sudo service nginx restart
+	# Restart
+	sudo service nginx restart
+fi
